pyproject.toml
[project]
name = "MuiceBot"
dynamic = ["version"]
description = "一个可以用来调用多种模型，且被用来练手的Nonebot项目"
readme = "README.md"
requires-python = ">=3.10, <=3.12"
dependencies = [
    "APScheduler>=3.11.0",
    "azure-ai-inference>=1.0.0b7",
    "dashscope>=1.22.1",
    "nonebot2>=2.4.1",
    "nonebot-adapter-console>=0.6.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-qq>=1.6.2",
    "nonebot-adapter-telegram>=0.1.0b20",
    "nonebot_plugin_alconna>=0.54.2",
    "nonebot_plugin_apscheduler>=0.5.0",
    "nonebot_plugin_localstore>=0.7.3",
    "nonebot_plugin_orm>=0.7.6",
    "ollama>=0.4.7",
    "openai>=1.64.0",
    "pydantic>=2.10.5",
    "httpx>=0.27.0",
    "ruamel.yaml>=0.18.10",
    "SQLAlchemy>=2.0.38",
    "websocket_client>=1.8.0",
    "setuptools-scm>=8.2.0"
]
authors = [
    { name = "Moemu", email = "i@snowy.moe" },
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.1.0",
    "nonebot-adapter-onebot>=2.4.6",
    "mypy>=1.15.0",
    "black>=25.1.0",
    "types-PyYAML",
]
test = [
    "nonebug>=0.4.3",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V12", module_name = "nonebot.adapters.onebot.v12" },
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "Telegram", module_name = "nonebot.adapters.telegram" },
    { name = "QQ", module_name = "nonebot.adapters.qq" },
]
plugins = ["nonebot_plugin_alconna", "nonebot_plugin_localstore", "nonebot_plugin_apscheduler"]
builtin_plugins = []

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"
tag_filter = "v*"
tag_regex = '^v(?:\D*)?(?P<version>([1-9][0-9]*!)?(0|[1-9][0-9]*)(\.(0|[1-9][0-9]*))*((a|b|c|rc)(0|[1-9][0-9]*))?(\.post(0|[1-9][0-9]*))?(\.dev(0|[1-9][0-9]*))?$)$'
fallback_version = "0.2.0"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"